Setup for BLOB test
